pyproject.toml
[project]
name = "pycfast"
version = "0.1.0"
description = "Python interface for building, running, and analyzing CFAST fire simulation models"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "WYGAS Benoît", email = "benoit.wgs@protonmail.com" }
]
maintainers = [
    { name = "WYGAS Benoît", email = "benoit.wgs@protonmail.com" }
]
keywords = ["fire", "simulation", "cfast", "modeling", "fire-safety", "engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11", 
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
]
dependencies = [
    "f90nml>=1.4.5",
    "pandas>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/bewygs/pycfast"
Documentation = "https://bewygs.github.io/pycfast"
Repository = "https://github.com/bewygs/pycfast"
Issues = "https://github.com/bewygs/pycfast/issues"
Changelog = "https://github.com/bewygs/pycfast/blob/main/CHANGELOG.md"

[project.optional-dependencies]
docs = [
    "pycfast[examples]",
    "ipython",
    "sphinx>=5.0",
    "myst-parser",
    "nbsphinx",
    "sphinx-gallery>=0.16",
    "linkify-it-py",
    "numpydoc>=1.9.0",
    "sphinx-copybutton>=0.5.2",
    "pydata-sphinx-theme>=0.16.1",
    "sphinx-autodoc-typehints>=3.0.1",
]
dev = [
    "ruff",
    "pytest",
    "mypy",
    "pre-commit",
    "coverage",
    "pytest-cov",
    "pandas-stubs",
]
examples = [
    "salib>=1.5.1",
    "scipy>=1.15.3",
    "matplotlib>=3.10.5",
    "jupyter>=1.1.1",
    "seaborn>=0.13.2",
    "scikit-learn>=1.7.1",
    "dask>=2025.7.0",
    "distributed>=2025.7.0",
    "bokeh>=3.8.0",
    "torch>=2.8.0",
]

[build-system]
requires = [
    "setuptools>=61.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "build",
    "dist",
    "__pycache__",
    "*.egg-info",
    "docs/", # conf.py
    "examples/**/*.ipynb" # exclude notebook for user readability
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "D",
]
ignore = [
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "DOC"
]
"examples/**" = [
    "B018",
    "E402",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D205",
    "D400",
    "DOC"
]


[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (may take several minutes)",
    "local: marks tests that use local verification data"
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
